% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mun1.R
\docType{data}
\name{data_mun1}
\alias{data_mun1}
\title{Municipal direct-democracy results — part 1}
\format{
A data frame.
}
\source{
Swiss Federal Statistical Office
}
\usage{
data_mun1
}
\description{
First part of the municipal direct-democracy dataset included
with RSwissPos.
}
\keyword{datasets}
