% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mun2.R
\docType{data}
\name{data_mun2}
\alias{data_mun2}
\title{Municipal direct-democracy results — part 2}
\format{
A data frame.
}
\source{
Swiss Federal Statistical Office
}
\usage{
data_mun2
}
\description{
Second part of the municipal direct-democracy dataset included
with RSwissPos.
}
\keyword{datasets}
