% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSwissvotes.R
\name{getSwissvotes}
\alias{getSwissvotes}
\title{Downloads and formats the latest version of the Swissvotes dataset.}
\format{
A data.frame object with 874 variables. The complete list of variables, their function and codes are listed in the Codebook available here: https://swissvotes.ch/storage/84d5514faa708639fc5586ab54b0c7aae6c56ae6b9485e653988dbf72bbb8784
}
\usage{
getSwissvotes(Column.names = "All")
}
\arguments{
\item{Column.names}{Indicates the Column name of interest that should be extracted from the data. Default value is 'All' which returns all the variables. It can take any number of existing column names from the swissvotes dataset.}
}
\value{
A dataset containing diverse information at the ballot level.
}
\description{
This function downloads the latest version of the Swissvotes dataset.
The data collects information at the ballot level since the first national ballot in 1848.
The codebook for the Swissvotes data set can be found here: https://swissvotes.ch/storage/84d5514faa708639fc5586ab54b0c7aae6c56ae6b9485e653988dbf72bbb8784
To cite, please use: Swissvotes (Year). Swissvotes, die Datenbank der eidgen\u00f6ssischen Volksabstimmungen.
For more information see: https://www.swissvotes.ch
}
\examples{
data_all <- getSwissvotes() ## Gets the Swissvotes data with all the variables

data_institutions <- getSwissvotes(c("anr", "rechtsform")) ### Get the direct democratic institutions and the id of the ballot.


}
