% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCap.R
\name{getCAP}
\alias{getCAP}
\title{Downloads and formats the latest version of the Comprative Agendas' project dataset on direct democracy.}
\format{
A data.frame object with 17 variables. The complete list of variables, their function and codes are listed in the Codebook available here: https://minio.la.utexas.edu/compagendas/codebookfiles/switzerland_direct_democracy_codebook.pdf
}
\usage{
getCAP(Column.names = "All")
}
\arguments{
\item{Column.names}{Indicates the Column name of interest that should be extracted from the data. Default value is 'All' which returns all the variables. It can take any number of existing column names from the Comparative Agenda's project data set.}
}
\value{
A dataset containing diverse information related to the .
}
\description{
This function downloads the latest version of the Comparative Agenda's project coding of Swiss direct democratic ballots.
The data collects information at the ballot level since the first national ballot in 1848.
The codebook for the Comparative Agenda's project data set can be found here: https://minio.la.utexas.edu/compagendas/codebookfiles/switzerland_direct_democracy_codebook.pdf
To cite, please use the following footnote: "The data used here were originally collected by Roy Gava, Pascal Sciarini, Anke Tresch and Frédéric Varone, with the support of the Swiss National Science Foundation (grant number 105511-119245/1 and project ‘The Mediatization of Political Decision Making’ sponsored as part of the National Center of Competence in Research ‘Challenges to Democracy in the 21st Century’). Neither SNSF nor the original collectors of the data bear any responsibility for the analysis reported here."
For more information see: https://www.comparativeagendas.net
}
\examples{
CAP <- getCAP() #### Gets all the variables
CAP_mal <- getCAP(c("anr", "majortopic")) ### Gets the identification number of the ballot and the majortopic


}
