% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCap.R
\name{getSwissGeo}
\alias{getSwissGeo}
\title{Download Swiss administrative boundaries}
\format{
A data.frame object with 17 variables. The complete list of variables, their function and codes are listed in the Codebook available here: https://minio.la.utexas.edu/compagendas/codebookfiles/switzerland_direct_democracy_codebook.pdf
}
\usage{
getSwissGeo <- function(
    year,
    type = "Municipality",
    detail = "K4",
    surface = "total",
    crs = 2056
)
}
\arguments{
\item{year}{Year of the administrative boundaries. The function retrieves the administrative state on January 1 of that year.}
\item{type}{Geographical unit to return. One of `"Municipality"`, `"District"`, `"Canton"`, `"Country"`, or `"Lake"`}
\item{detail}{Level of geographical detail. Either `"K4"` or `"G1"`. `"K4"` is the default and is recommended for maps of Switzerland.}
\item{surface}{Surface representation. Either `"total"` or `"vegetation"`. Vegetation surfaces are only available with `detail = "K4"` for municipalities, districts, and cantons.}
}

\value{
An object of class `sf` containing diverse information including administrative boudaries of various surfaces.
}
\description{
Downloads historical administrative boundaries from the Swiss Federal Statistical Office (BFS).

Boundaries are available for municipalities, districts, cantons, Switzerland, and lakes for historical states since 1850.
}
\examples{
municipalities <- getSwissGeo(2025)

 municipalities_vf <- getSwissGeo(
   2025,
   surface = "vegetation"
 )

 lakes <- getSwissGeo(
   2025,
   type = "Lake"
 )

}
